requirements.txt
fastapi
uvicorn
python-dotenv
requests
tenacity
pandas
pydantic
python-multipart
